refactor imports for better performance * move all imports to format-specific submodules * refactor tests accordingly * remove tests for DeepLabCut (was causing problems for me) * make interface_list which is used by the yaml workflow. This can be improved later * modify jsonschema requirement. This was needed for me. Maybe it's a new thing.
