+++ requirements-full.txt
@@ -10,7 +10,7 @@ hdf5storage==0.1.18
 scipy>=1.4.1
 h5py>=2.10.0
 PyYAML==5.4
-jsonschema==3.2.0
+jsonschema[format]==4.9.2
 psutil==5.8.0
 lxml==4.9.1
 opencv-python==4.5.1.48
