+++ src/neuroconv/datainterfaces/behavior/deeplabcut/__init__.py
@@ -1,0 +1,1 @@
+from .deeplabcutdatainterface import DeepLabCutInterface
