+++ src/neuroconv/datainterfaces/behavior/movie/__init__.py
@@ -1,0 +1,1 @@
+from .moviedatainterface import MovieInterface
