+++ src/neuroconv/datainterfaces/ecephys/blackrock/__init__.py
@@ -1,0 +1,4 @@
+from .blackrockdatainterface import (
+    BlackrockRecordingExtractorInterface,
+    BlackrockSortingExtractorInterface,
+)
