+++ src/neuroconv/datainterfaces/ecephys/ced/__init__.py
@@ -1,0 +1,1 @@
+from .ceddatainterface import CEDRecordingInterface
