+++ src/neuroconv/datainterfaces/ecephys/cellexplorer/__init__.py
@@ -1,0 +1,1 @@
+from .cellexplorerdatainterface import CellExplorerSortingInterface
