+++ src/neuroconv/datainterfaces/ecephys/edf/__init__.py
@@ -1,0 +1,1 @@
+from .edfdatainterface import EDFRecordingInterface
