+++ src/neuroconv/datainterfaces/ecephys/intan/__init__.py
@@ -1,0 +1,1 @@
+from .intandatainterface import IntanRecordingInterface
