+++ src/neuroconv/datainterfaces/ecephys/kilosort/__init__.py
@@ -1,0 +1,1 @@
+from .kilosortdatainterface import KilosortSortingInterface
