+++ src/neuroconv/datainterfaces/ecephys/maxwell/__init__.py
@@ -1,0 +1,1 @@
+from .maxwelldatainterface import MaxwellRecordingInterface
