+++ src/neuroconv/datainterfaces/ecephys/neuralynx/__init__.py
@@ -1,0 +1,4 @@
+from .neuralynxdatainterface import (
+    NeuralynxSortingInterface,
+    NeuralynxRecordingInterface,
+)
