+++ src/neuroconv/datainterfaces/ecephys/neuroscope/__init__.py
@@ -1,0 +1,6 @@
+from .neuroscopedatainterface import (
+    NeuroscopeSortingInterface,
+    NeuroscopeRecordingInterface,
+    NeuroscopeLFPInterface,
+    NeuroscopeMultiRecordingTimeInterface,
+)
