+++ src/neuroconv/datainterfaces/ecephys/openephys/__init__.py
@@ -1,0 +1,4 @@
+from .openephysdatainterface import (
+    OpenEphysRecordingExtractorInterface,
+    OpenEphysSortingExtractorInterface,
+)
