+++ src/neuroconv/datainterfaces/ecephys/phy/__init__.py
@@ -1,0 +1,1 @@
+from .phydatainterface import PhySortingInterface
