+++ src/neuroconv/datainterfaces/ecephys/spikegadgets/__init__.py
@@ -1,0 +1,1 @@
+from .spikegadgetsdatainterface import SpikeGadgetsRecordingInterface
