+++ src/neuroconv/datainterfaces/ecephys/spikeglx/__init__.py
@@ -1,0 +1,4 @@
+from .spikeglxdatainterface import (
+    SpikeGLXLFPInterface,
+    SpikeGLXRecordingInterface,
+)
