+++ src/neuroconv/datainterfaces/ecephys/tutorial/__init__.py
@@ -1,0 +1,2 @@
+from .recordingtutorialdatainterface import RecordingTutorialInterface
+from .sortingtutorialdatainterface import SortingTutorialInterface
