+++ src/neuroconv/datainterfaces/ophys/caiman/__init__.py
@@ -1,0 +1,1 @@
+from .caimandatainterface import CaimanSegmentationInterface
