+++ src/neuroconv/datainterfaces/ophys/cnmfe/__init__.py
@@ -1,0 +1,1 @@
+from .cnmfedatainterface import CnmfeSegmentationInterface
