+++ src/neuroconv/datainterfaces/ophys/extract/__init__.py
@@ -1,0 +1,1 @@
+from .extractdatainterface import ExtractSegmentationInterface
