+++ src/neuroconv/datainterfaces/ophys/hdf5/__init__.py
@@ -1,0 +1,1 @@
+from .hdf5datainterface import Hdf5ImagingInterface
