+++ src/neuroconv/datainterfaces/ophys/sbx/__init__.py
@@ -1,0 +1,1 @@
+from .sbxdatainterface import SbxImagingInterface
