+++ src/neuroconv/datainterfaces/ophys/scanimage/__init__.py
@@ -1,0 +1,1 @@
+from .scanimageimaginginterface import ScanImageImagingInterface
