+++ src/neuroconv/datainterfaces/ophys/sima/__init__.py
@@ -1,0 +1,1 @@
+from .simadatainterface import SimaSegmentationInterface
