+++ src/neuroconv/datainterfaces/ophys/suite2p/__init__.py
@@ -1,0 +1,1 @@
+from .suite2pdatainterface import Suite2pSegmentationInterface
