+++ src/neuroconv/datainterfaces/ophys/tiff/__init__.py
@@ -1,0 +1,1 @@
+from .tiffdatainterface import TiffImagingInterface
